Fix memory bugs and replace WriteLock with pthread_rwlock for thread safety
- VectorInt: Replace malloc+memcpy with memmove in inserto/removefrom (fixes
  zero-size malloc bug when removing last element, eliminates unnecessary
  memory allocation)
- StringList: Fix double-free in String_Split (free(copy) freed string_cpy's
  internal buffer; use free_String instead)
- Hashtable: Fix signed/unsigned comparison in insertid_intohashtable
- XmlDB: Replace bool WriteLock with pthread_rwlock_t for proper thread
  synchronization (atomic operations, memory barriers, CPU-efficient waiting
  instead of busy spin)
- Initialize rwlock in init_Database, destroy in free_DB
- Link pthread library in CMakeLists.txt

Agent-Logs-Url: https://github.com/LIJUCHACKO/XmlDB_C/sessions/2362770a-e5d7-49a7-8a4f-9fa65aa36edb

+++ xmldb_lib/Hashtable.c
@@ -37,7 +37,7 @@ void insertid_intohashtable(struct Hashtable* HashTB,int hashno ,int  nodeId,int
         return;
     }
     int index = Vect->length - 1;
-    unsigned long lastno=Vect->length;
+    int lastno=(int)Vect->length;
     while(nodeNoToLineno->items[HashTB->lists[hashno].items[index]] > reference_linenotoinsert ){
         lastno=index;
         index--;
@@ -45,7 +45,7 @@ void insertid_intohashtable(struct Hashtable* HashTB,int hashno ,int  nodeId,int
             break;
         }
     }
-    if(lastno > Vect->length - 1){
+    if(lastno > (int)Vect->length - 1){
         appendto_VectorInt(Vect,nodeId);
         return;
     }
